requirements.txt
coreapi==2.3.3
Django==2.2.7
django-coreapi==2.3.0
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
drf-yasg==1.17.0
model-bakery==1.0.2